refactor: Update .gitlab-ci.yml
Simplify the image tag in the .gitlab-ci.yml file by removing the unnecessary "slim" tag. This ensures that the latest LTS version of Node.js is used for the CI pipeline.

Also, update the cache paths to include the .pnpm-store directory and the .cache directory.

Lastly, add a new script to install Chromium for Playwright before running the build process.