refactor: Update .gitlab-ci.yml to use Microsoft Playwright image
This commit updates the .gitlab-ci.yml file to use the Microsoft Playwright image (mcr.microsoft.com/playwright:v1.47.2-noble) instead of the node:lts image. It removes the installation of Chromium for Playwright in the "deps:install" and "pages" stages since it is now included in the Microsoft Playwright image. This change ensures consistent and up-to-date dependencies for Playwright across the CI pipeline.